refactor(Board): 보드 초기화 로직을 별도의 factory 클래스로 분리
