+++ src/main/java/chess/domain/ChessGame.java
@@ -1,5 +1,6 @@
 package chess.domain;
 
+import chess.domain.board.Board;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
