+++ src/test/java/chess/domain/chessGame/BlackPawnTest.java
@@ -1,6 +1,5 @@
 package chess.domain.chessGame;
 
-import chess.domain.Board;
 import chess.domain.ChessGame;
 import chess.domain.Color;
 import chess.domain.Column;
@@ -8,6 +7,7 @@ import chess.domain.Piece;
 import chess.domain.PieceType;
 import chess.domain.Position;
 import chess.domain.Row;
+import chess.domain.board.Board;
 import java.util.List;
 import java.util.Map;
 import org.assertj.core.api.Assertions;
